Refactor plural functions into their own module
