+++ config/nunjucks/filters.js
@@ -33,40 +33,14 @@ const {
 } = require('lodash')
 require('numeral/locales/en-gb')
 
+const { pluralise } = require('../../src/lib/plural')
+
 numeral.locale('en-gb')
 
 const { longDateFormat, mediumDateTimeFormat, currencyFormat } = require('../../config')
 
 function isNotEmpty (value) {
   return !isNil(value) && !/^\s*$/.test(value) && !(isPlainObject(value) && isEmpty(value))
-}
-
-function pluralise (string, count, pluralisedWord) {
-  if (parseInt(count, 10) !== 1) {
-    if (pluralisedWord) {
-      string = pluralisedWord
-    } else if (string.match(/[^aeiou]y$/)) {
-      string = string.replace(/y$/, 'ies')
-    } else {
-      string += 's'
-    }
-  }
-
-  return string
-}
-
-function singularise (string, count, singularisedWord) {
-  if (parseInt(count, 10) === 1) {
-    if (singularisedWord) {
-      return singularisedWord
-    } else if (string.endsWith('ies')) {
-      return string.substr(0, string.length - 3) + 'y'
-    } else {
-      return string.substr(0, string.length - 1)
-    }
-  }
-
-  return string
 }
 
 const filters = {
@@ -89,7 +63,6 @@ const filters = {
   isPlainObject,
   isString,
   pluralise,
-  singularise,
   sentenceCase: Case.sentence,
 
   encodeQueryString (value) {
